Remove unnecessary checkpointCallStacks
The purpose of checkpointCallStack is to add the CallStack to any
currently thrown exception i.e. we want this if we call a function that
might throw an exception _and does not include the CallStack_. Crucially,
we do _not_ want to add it to arbitrary functions because it is unnecessary
if the downstream functions already include the CallStack, and it adds significant
noise.

Thus we do not add it on functions that operate over some typeclass, as that penalizes
instances that are already doing the right thing.