Rename STM helper fns *M -> *A
The STM functions have been renamed e.g. from

newTVarM :: MonadSTM m => a -> m (TVar a)

to

newTVarA :: MonadSTM m => a -> m (TVar a)

This is due to the existence of _other_ functions we may want to lift
as well:

newTVarIO :: a -> IO (TVar a)

This *IO series of functions exists as an alternative to atomically
for various reasons (e.g. performance, unsafePerformIO). If we ever
want to lift these, then a natural candidate is:

newTVarM :: MonadTVar m => a -> m (TVar a)

But this would clash with our existing newTVarM, thus we rename the
latter to newTVarA, for "atomically".