Add new TargetName type for directory copying
Previously, if one wanted to copy a directory in a 'top-level' manner,
they would have to pass 'targetName = Just ' i.e. a blank target.
This works fine, but the discoverability / usage is fairly low, so
we introduce an explicit type that can capture this usage.

We also add tests to verify that 'merged copies' works.