Create test data directory in tests
Our effects-fs unit tests rely on a data directory. Originally we had
this data directory committed, but unfortunately this causes a stack
error. In particular, stack dies when copying the effects-fs repo when
it is a dependency, due to a bad sym link (which a particular test
needs!)

Thus we instead create the data directory within the tests themselves,
fixing the copy issue.