Add unix effect, effects-fs-utils lib
We add an effect for unix, in addition to unix-compat. This leaves us
in an awkward position, because now we have two PathType types
(defined in both). Moreover, effects-fs depends on effects-unix-compat
for this type, but really this type should be shared.

Thus we split out PathType (and other common FileSystem utils) into
a new utility package: effects-fs-utils. Then, effects-fs,
effects-unix-compat, and effects-unix can all depend on this package.