Rename effects-logger-ns -> effects-logger
The namespace addition really doesn't need to be in a separate lib.
A separate module is more ergonomic.

We also change the namespace so it is no longer its own effect. It's
merely MonadReader on an abstract Env type w/ an optic field.
Eventually it would be nice to use HasField.